refactor: share registry-location helpers between CLI and Ansible module
canasta.py re-implemented get_config_dir / conf.json reading / find-by-path
that already lived in the canasta_registry Ansible module, kept in sync by
hand (its get_config_dir even carried a "Mirrors canasta_registry.py logic"
comment). The two copies had diverged: canasta.py detected root with
os.geteuid()==0 while the module used a name-based is_root(), so they could
disagree about where conf.json lives.

Extract the shared logic into roles/common/module_utils/canasta_config.py
(is_root, get_config_dir, config_path, read_config, write_config,
find_by_path). canasta_registry imports and re-exports it; canasta.py
imports it off the same module_utils directory. The registry-location and
lookup logic now lives in exactly one place.

Tests for the helpers move to test_canasta_config.py; test_canasta_registry
keeps a re-export smoke check plus the module-specific coverage.

Addresses item 3 of #773.
