+++ roles/common/library/canasta_registry.py
@@ -79,78 +79,26 @@ options:
 
 import json
 import os
-import platform
-import pwd
 
 from ansible.module_utils.basic import AnsibleModule
 
+# Registry location and lookup helpers are shared with the canasta.py CLI
+# wrapper so the two never disagree about where conf.json lives.
+from ansible.module_utils.canasta_config import (
+    config_path,
+    find_by_path,
+    get_config_dir,
+    is_root,
+    read_config,
+    write_config,
+)
 
-CONFIG_FILENAME = "conf.json"
-
-
-def get_config_dir(override=None):
-    """Determine the config directory using the platform's conventional
-    per-user config location:
-    - macOS: ~/Library/Application Support/canasta
-    - Linux: $XDG_CONFIG_HOME/canasta or ~/.config/canasta
-    """
-    if override:
-        return override
-    env_dir = os.environ.get("CANASTA_CONFIG_DIR")
-    if env_dir:
-        return env_dir
-    if is_root():
-        return "/etc/canasta"
-    if platform.system() == "Darwin":
-        config_home = os.path.join(
-            os.path.expanduser("~"), "Library", "Application Support"
-        )
-    else:
-        config_home = os.environ.get(
-            "XDG_CONFIG_HOME",
-            os.path.join(os.path.expanduser("~"), ".config"),
-        )
-    return os.path.join(config_home, "canasta")
-
-
-def is_root():
-    """Check if running as root."""
-    try:
-        return pwd.getpwuid(os.getuid()).pw_name == "root"
-    except (KeyError, AttributeError):
-        return os.getuid() == 0
-
-
-def config_path(config_dir):
-    """Return the full path to conf.json."""
-    return os.path.join(config_dir, CONFIG_FILENAME)
-
-
-def read_config(config_dir):
-    """Read and return the registry config, creating it if needed."""
-    path = config_path(config_dir)
-    if not os.path.exists(path):
-        return {"Instances": {}}
-    with open(path, "r") as f:
-        content = f.read().strip()
-    if not content:
-        return {"Instances": {}}
-    data = json.loads(content)
-    # Migrate legacy "Installations" key
-    if "Installations" in data and "Instances" not in data:
-        data["Instances"] = data.pop("Installations")
-    if "Instances" not in data:
-        data["Instances"] = {}
-    return data
-
-
-def write_config(config_dir, data):
-    """Write the registry config to conf.json."""
-    os.makedirs(config_dir, mode=0o755, exist_ok=True)
-    path = config_path(config_dir)
-    with open(path, "w") as f:
-        json.dump(data, f, indent=4)
-        f.write("\n")
+# Re-exported for backwards compatibility with importers/tests that
+# referenced canasta_registry.* before the helpers were extracted.
+__all__ = [
+    "config_path", "find_by_path", "get_config_dir", "is_root",
+    "read_config", "write_config", "instance_to_dict", "run_module",
+]
 
 
 def instance_to_dict(instance):
@@ -175,20 +123,6 @@ def instance_to_dict(instance):
     if instance.get("dockerHost"):
         result["dockerHost"] = instance["dockerHost"]
     return result
-
-
-def find_by_path(instances, search_path):
-    """Walk up from search_path to find a matching instance."""
-    search_path = os.path.abspath(search_path)
-    while True:
-        for inst_id, inst in instances.items():
-            if os.path.abspath(inst.get("path", "")) == search_path:
-                return inst_id, inst
-        parent = os.path.dirname(search_path)
-        if parent == search_path:
-            break
-        search_path = parent
-    return None, None
 
 
 def run_module():
