+++ tests/unit/conftest.py
@@ -20,7 +20,7 @@ sys.path.insert(0, os.path.join(ROLES_DIR, "extensions_skins", "library"))
 import importlib  # noqa: E402
 _module_utils_dir = os.path.join(ROLES_DIR, "common", "module_utils")
 sys.path.insert(0, _module_utils_dir)
-for _name in ("canasta_validate",):
+for _name in ("canasta_validate", "canasta_config"):
     _mod = importlib.import_module(_name)
     sys.modules.setdefault("ansible.module_utils.%s" % _name, _mod)
